% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_feature_data.R
\name{add_taxonomy}
\alias{add_taxonomy}
\title{Add a taxonomy table to qSIP abundance data}
\usage{
add_taxonomy(x, ...)
}
\arguments{
\item{x}{An object of \code{qsip_feature_data} class}

\item{taxa}{A taxa table}

\item{id}{The column name for the taxa ids that match the ids in the
abundance table}
}
\description{
Add a taxonomy table to qSIP abundance data
}
\keyword{abundance}
