% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_EAF}
\alias{calculate_EAF}
\title{Calculate EAF value (internal)}
\usage{
calculate_EAF(M_labeled, M, M_labeledmax, isotope)
}
\arguments{
\item{M_labeled}{Molecular weight of the labeled feature}

\item{M}{Molecular weight of the unlabeled feature}

\item{M_labeledmax}{Theoretical molecular weight of a completely labeled feature}

\item{isotope}{The isotope to determine which calculation to run}
}
\value{
EAF value
}
\description{
Calculate EAF value (internal)
}
\keyword{internal}
