% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calculate_filter_efficiency}
\alias{calculate_filter_efficiency}
\title{Plot features vs abundance with labeled/unlabeled/retained categories}
\usage{
calculate_filter_efficiency(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A filtered qsip_data object (or list)}
}
\value{
A ggplot2 object
}
\description{
Plot features vs abundance with labeled/unlabeled/retained categories
}
