% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_gc_from_density}
\alias{calculate_gc_from_density}
\title{Calculate GC\% from density/WAD}
\usage{
calculate_gc_from_density(density, method = "MM")
}
\arguments{
\item{density}{(\emph{numeric}) Density or WAD values}

\item{method}{(\emph{string, default: MM}) The GC\% calculation method}
}
\value{
A vector of GC\% values
}
\description{
This function takes a \code{density} value and an optional \code{method} and returns the
predicted GC\% content of a DNA sequence with that density.
}
\details{
The \code{method} parameter changes the formula from that provided by McHugh &
Morrissey (\code{MM}, unpublished) or Schildkraut (\code{S}, 1962).

This function corresponds to equation 5 from Hungate, 2015
}
