% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_gc_from_sequence}
\alias{calculate_gc_from_sequence}
\title{Calculate GC\% from a sequence}
\usage{
calculate_gc_from_sequence(sequence)
}
\arguments{
\item{sequence}{(\emph{string}) Density or WAD values}
}
\value{
A vector of GC\% values
}
\description{
This function takes a sequence and counts the number of G and C bases, and divides
by the total sequence length.
}
