% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_M}
\alias{calculate_M}
\title{Calculate molecular weight of the labeled feature (internal)}
\usage{
calculate_M(G)
}
\arguments{
\item{G}{(\emph{numeric}) GC content of a feature, ranges from 0-1.}
}
\value{
\code{M} is the molecular weight of a sequence with \code{G} GC content
}
\description{
The GC value given to this equation is usually calculated from the density
value, not derived from the sequence itself.
}
\details{
This function corresponds to equation 6 from Hungate, 2015.
}
