% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_M_heavy}
\alias{calculate_M_heavy}
\title{Calculate M_heavy}
\usage{
calculate_M_heavy(propO, M)
}
\arguments{
\item{propO}{(\emph{numeric}) The proportion of oxygen coming from 18H2O versus other sources}

\item{M}{(\emph{numeric}) The mass of the molecule}
}
\description{
This is equation 4 from Koch, 2018
}
\keyword{internal}
