% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.R
\name{calculate_M_labeled}
\alias{calculate_M_labeled}
\title{Calculate molecular weight of the labeled feature (internal)}
\usage{
calculate_M_labeled(M, W_lab_mean, W_unlab_mean)
}
\arguments{
\item{M}{(\emph{numeric}) Molecular weight of the unlabeled feature}

\item{W_lab_mean}{(\emph{numeric}) WAD (or mean WAD) value of the labeled feature}

\item{W_unlab_mean}{(\emph{numeric}) WAD (or mean WAD) value of the unlabeled feature}
}
\value{
\code{M_labeled} is the molecular weight of the labeled feature
}
\description{
Calculate molecular weight of the labeled feature (internal)
}
