% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{calculate_N_light_it}
\alias{calculate_N_light_it}
\title{Calculate N_Light_it}
\usage{
calculate_N_light_it(N_total_it, M_heavy, M_labeled, M)
}
\description{
This is equation 3 from Koch, 2018
}
