% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_resampled_wads.R
\name{calculate_resampled_wads}
\alias{calculate_resampled_wads}
\title{Internal function for resampling WAD values}
\usage{
calculate_resampled_wads(i, wad_data, type)
}
\arguments{
\item{i}{(\emph{integer}) The specific iteration of the resampling}

\item{wad_data}{(\emph{dataframe}) A WAD dataframe to resample columns from}

\item{type}{(\emph{string}) Text for whether the wad data is from labeled or unlabeled data}
}
\value{
The resampling data that will be boot in \verb{@resamples}
}
\description{
Designed as a purrr::map() function called by run_resampling.R
}
