% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_source_wads}
\alias{calculate_source_wads}
\title{Calculate global weighted average density (WAD) value for a source_mat_id (internal)}
\usage{
calculate_source_wads(sample_data)
}
\arguments{
\item{sample_data}{(\emph{qsip_sample_data}) Sample data object}
}
\value{
A dataframe with two columns, 1) the source_mat_id and 2) the global
WAD value for that source_mat_id
}
\description{
Calculate global weighted average density (WAD) value for a source_mat_id (internal)
}
