% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{calculate_time_zero_totals}
\alias{calculate_time_zero_totals}
\title{Calculate time zero feature totals}
\usage{
calculate_time_zero_totals(qsip_data_object, time = "time", value = 0)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{time}{(\emph{character}) The name of the time column in the source data}

\item{value}{(\emph{numeric}) The value of the time column to filter on}
}
\description{
Calculate time zero feature totals
}
