% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tube_rel_abundance.R
\name{calculate_tube_rel_abundance}
\alias{calculate_tube_rel_abundance}
\title{Calculate tube relative abundance}
\usage{
calculate_tube_rel_abundance(source_data, sample_data, feature_data)
}
\arguments{
\item{source_data}{(\emph{qsip_source_data}) A qSIP source data object}

\item{sample_data}{(\emph{qsip_sample_data}) A qSIP sample data object}

\item{feature_data}{(\emph{qsip_feature_data}) A qSIP feature data object}
}
\value{
A long format dataframe with one row per \code{feature_id} per \code{sample_id}
}
\description{
The "tube level relative abundance" has two layers of relative abundance. It first
takes the \code{raw_abundance} counts per sample, and divides by the total to get
the \code{rel_abundance} values. Next, it takes these values and divides them by the
\code{gradient_pos_rel_amt} values stored at the source level which normalizes the \code{rel_abundance}
values by the amount of sample per fraction. This is the final value stored in
the \code{tube_rel_abundance} column.
}
\details{
To speed up the calculations, this function removes \code{feature_ids} in a sample
if the abundance is zero.

v0.10.3 updated this function to have different behavior for feature type 'normalized'.
This type should use the raw data in the feature table, putting those values into the
\code{tube_rel_abundance} column.
}
