% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wads.R
\name{calculate_wads}
\alias{calculate_wads}
\title{Calculate weighted average density (WAD) values}
\usage{
calculate_wads(tube_rel_abundance)
}
\arguments{
\item{tube_rel_abundance}{(\emph{dataframe}) Output from \code{calculate_tube_rel_abundance()}}
}
\value{
An list with two objects, 1) a dataframe with WAD info for the feature_ids
found in at least one sample, and 2) the fraction counts for all feature_ids,
including those not found at all in some samples. It also prints a message to
the screen with a count of feature_ids that are entirely missing from at least
one sample.
}
\description{
Calculate weighted average density (WAD) values
}
