% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_Z}
\alias{calculate_Z}
\title{Calculate the WAD difference Z (internal)}
\usage{
calculate_Z(labeled, unlabeled)
}
\arguments{
\item{labeled}{(\emph{string}) The column with the labeled WAD or mean labeled WAD value}

\item{unlabeled}{(\emph{string}) The column with the unlabeled WAD or mean unlabeled WAD value}
}
\value{
a value for the difference between labeled and unlabeled
}
\description{
This is just basic subtraction, so probably overkill to make a function!
}
\details{
This function corresponds to equation 4 from Hungate, 2015
}
\keyword{internal}
