% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completely_labeled_values.R
\name{completely_labeled_values}
\alias{completely_labeled_values}
\title{Returns the max labeling of a given isotope}
\usage{
completely_labeled_values(isotope)
}
\arguments{
\item{isotope}{The heavy isotope (13C, 15N or 18O)}
}
\description{
Returns the max labeling of a given isotope
}
