% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{correct_gpd_rlm}
\alias{correct_gpd_rlm}
\title{Correct gradient position density (gpd) values using linear regression approach}
\usage{
correct_gpd_rlm(
  qsip_data_object,
  method = "siegel",
  fraction_cutoff = 5,
  source_cutoff = 3,
  return = "qsip_data_object"
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}

\item{method}{(\emph{character}) The lm method to use}
}
\description{
Correct gradient position density (gpd) values using linear regression approach
}
