% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_feature_df}
\alias{example_feature_df}
\title{Example Feature Abundance Dataframe}
\format{
\subsection{\code{example_feature_df}}{

A data frame with 2,030 rows and 285 columns:
\describe{
\item{ASV}{The unique feature ID}
\item{Columns 2-285}{IDs for all samples}
}
}
}
\usage{
example_feature_df
}
\description{
An example feature table with abundances
}
\seealso{
Other "example datasets":
\code{\link{example_feature_object}},
\code{\link{example_group_dataframe}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_growth_t0}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_df}},
\code{\link{example_sample_object}},
\code{\link{example_source_df}},
\code{\link{example_source_object}}
}
\concept{"example datasets"}
\keyword{datasets}
