% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_group_dataframe}
\alias{example_group_dataframe}
\title{Example data frame for run_comparison_groups()}
\format{
\subsection{\code{example_group_dataframe}}{

a 6x10 tibble
}
}
\usage{
example_group_dataframe
}
\description{
An dataframe with 6 example comparisons.
}
\seealso{
Other "example datasets":
\code{\link{example_feature_df}},
\code{\link{example_feature_object}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_growth_t0}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_df}},
\code{\link{example_sample_object}},
\code{\link{example_source_df}},
\code{\link{example_source_object}}
}
\concept{"example datasets"}
\keyword{datasets}
