% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsip_feature_data}
\alias{example_qsip_feature_data}
\title{Example qSIP Feature Abundance Data}
\format{
\subsection{\code{example_qsip_feature_data}}{

A data frame with 2,030 rows and 285 columns:
\describe{
\item{ASV}{The unique feature ID}
\item{Columns 2-285}{IDs for all samples}
}
}
}
\usage{
example_qsip_feature_data
}
\description{
An example feature table with abundances
}
\seealso{
Other "example datasets": 
\code{\link{example_qsip_sample_data}},
\code{\link{example_qsip_source_data}}
}
\concept{"example datasets"}
\keyword{datasets}
