% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsip_growth_t0}
\alias{example_qsip_growth_t0}
\title{Example qSIP Time Zero Growth Dataframe}
\format{
\subsection{\code{example_qsip_growth_t0}}{

A 364x3 tibble
}
}
\usage{
example_qsip_growth_t0
}
\description{
An example dataframe with time zero abundances for growth calculations
}
\seealso{
Other "example datasets":
\code{\link{example_feature_df}},
\code{\link{example_feature_object}},
\code{\link{example_group_dataframe}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_df}},
\code{\link{example_sample_object}},
\code{\link{example_source_df}},
\code{\link{example_source_object}}
}
\concept{"example datasets"}
\keyword{datasets}
