% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsip_sample_data}
\alias{example_qsip_sample_data}
\title{Example qSIP Sample Data}
\format{
\subsection{\code{example_qsip_sample_data}}{

A data frame with 284 rows and 6 columns:
\describe{
\item{sample}{The unique sample id of the fraction}
\item{source}{The parent source id for this sample}
\item{Fraction}{The gradient position}
\item{density_g_ml}{The density of this fraction}
\item{dna_conc}{The concentration of DNA post-fractionation}
\item{avg_16S_g_soil}{qPCR amplicon count}
}
}
}
\usage{
example_qsip_sample_data
}
\description{
An example sample data table with fraction-related metadata
}
\seealso{
Other "example datasets": 
\code{\link{example_qsip_feature_data}},
\code{\link{example_qsip_source_data}}
}
\concept{"example datasets"}
\keyword{datasets}
