% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsip_source_data}
\alias{example_qsip_source_data}
\title{Example qSIP Source Data}
\format{
\subsection{\code{example_qsip_source_data}}{

A data frame with 15 rows and 5 columns:
\describe{
\item{source}{The unique name of the biological replicate or source_mat_id}
\item{total_copies_per_g}{The total amplicon copies as determined by qPCR}
\item{total_dna}{The total DNA isolated from the source}
\item{Isotope}{The isotope used in this source}
\item{Moisture}{An example of a relevant treatment}
}
}
}
\usage{
example_qsip_source_data
}
\description{
An example source data table with top-level (pre-fractionation) metadata
}
\seealso{
Other "Example datasets": 
\code{\link{example_qsip_feature_data}},
\code{\link{example_qsip_sample_data}}
}
\concept{"Example datasets"}
\keyword{datasets}
