% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_source_df}
\alias{example_source_df}
\title{Example Source Dataframe}
\format{
\subsection{\code{example_source_df}}{

A data frame with 15 rows and 5 columns:
\describe{
\item{source}{The unique name of the biological replicate or source_mat_id}
\item{total_copies_per_g}{The total amplicon copies as determined by qPCR}
\item{total_dna}{The total DNA isolated from the source}
\item{Isotope}{The isotope used in this source}
\item{Moisture}{An example of a relevant treatment}
}
}
}
\usage{
example_source_df
}
\description{
An example source data table with top-level (pre-fractionation) metadata
}
\seealso{
Other "example datasets":
\code{\link{example_feature_df}},
\code{\link{example_feature_object}},
\code{\link{example_group_dataframe}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_growth_t0}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_df}},
\code{\link{example_sample_object}},
\code{\link{example_source_object}}
}
\concept{"example datasets"}
\keyword{datasets}
