% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_source_object}
\alias{example_source_object}
\title{Example qSIP Source Object}
\format{
\subsection{\code{example_source_object}}{

a qsip_source_data object
}
}
\usage{
example_source_object
}
\description{
An example qsip_source_data object
}
\seealso{
Other "example datasets":
\code{\link{example_feature_df}},
\code{\link{example_feature_object}},
\code{\link{example_group_dataframe}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_growth_t0}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_df}},
\code{\link{example_sample_object}},
\code{\link{example_source_df}}
}
\concept{"example datasets"}
\keyword{datasets}
