% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{fraction_results_message}
\alias{fraction_results_message}
\title{Filter features by fraction message formatting}
\usage{
fraction_results_message(by_fraction)
}
\arguments{
\item{by_fraction}{by_fraction dataframe from filter_features}
}
\description{
Filter features by fraction message formatting
}
