% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{get_correction_compression}
\alias{get_correction_compression}
\title{Get compression values from a rlm correction}
\usage{
get_correction_compression(corrections)
}
\arguments{
\item{corrections}{(\emph{tibble}) A tibble of corrections from \code{correct_gpd_rlm()}}
}
\description{
Get compression values from a rlm correction
}
