% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_objects.R
\name{get_dataframe}
\alias{get_dataframe}
\title{Return the original dataframe from a qsip_feature_data object}
\usage{
## S7 method for class <qSIP2::qsip_feature_data>
get_dataframe(x, original_headers = FALSE)
}
\arguments{
\item{x}{(\emph{qsip_feature_data}) A qSIP feature data object}

\item{original_headers}{(\emph{logical}) Return the original column names}

\item{type}{(\emph{source}, \emph{sample} or \emph{feature}) When giving a qSIP object the type determines which dataframe to retrieve}
}
\description{
Return the original dataframe from a qsip_feature_data object
}
