% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{get_delta_distributions}
\alias{get_delta_distributions}
\title{Get the delta distribution for the contrasts from resampling data}
\usage{
get_delta_distributions(resampled_EAF, contrasts)
}
\arguments{
\item{resampled_EAF}{a tibble}

\item{contrasts}{A validated contrasts table}
}
\value{
a list of deltas and messages
}
\description{
Get the delta distribution for the contrasts from resampling data
}
\keyword{internal}
