% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{get_delta_stats}
\alias{get_delta_stats}
\title{Calculate the delta EAF statistics}
\usage{
get_delta_stats(delta_distributions, confidence = 0.95)
}
\arguments{
\item{delta_distributions}{a tibble}

\item{confidence}{(numeric, default: 0.95) confidence interval}
}
\value{
a tibble of statistics
}
\description{
Calculate the delta EAF statistics
}
\keyword{internal}
