% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_data.R
\name{get_feature_data}
\alias{get_feature_data}
\title{Get feature dataframe from qSIP2 object}
\usage{
get_feature_data(feature_data, original_headers = FALSE)
}
\arguments{
\item{feature_data}{(\emph{qsip_feature_data or qsip_data}) A qSIP2 object with feature data}

\item{original_headers}{(\emph{boolean, default: FALSE}) If TRUE, the original column names will be used}
}
\value{
A tibble with the feature data
}
\description{
This function will return the feature data stored in a qSIP2 object as a tibble.
}
