% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feature_ids}
\alias{get_feature_ids}
\title{Return the feature_ids in a qsip object}
\usage{
get_feature_ids(qsip_data_object, filtered = FALSE)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} object that has been run through \code{run_resampling()}}

\item{filtered}{(\emph{Boolean}) If TRUE, return the feature_ids from the filtered data}
}
\description{
Return the feature_ids in a qsip object
}
