% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_filter_results}
\alias{get_filter_results}
\title{Tabular output summarizing filtering}
\usage{
get_filter_results(qsip_data_object, type = "counts")
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class that has been filtered}

\item{type}{(\emph{string, default: counts}) The type of output to return. Either "counts" or "feature_ids"}
}
\description{
Tabular output summarizing filtering
}
