% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_feature_summary.R
\name{get_filtered_feature_summary}
\alias{get_filtered_feature_summary}
\title{Return filtering info for a specific feature ID}
\usage{
get_filtered_feature_summary(qsip_data_object, feature_id)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object that has been filtered}

\item{feature_id}{(\emph{string}) A specific feature_id}
}
\description{
Return filtering info for a specific feature ID
}
