% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_filtered_source_counts}
\alias{get_filtered_source_counts}
\title{Return the number of sources the feature is found in}
\usage{
get_filtered_source_counts(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object that has been filtered}
}
\description{
Return the number of sources the feature is found in
}
