% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_growth_data.R
\name{get_growth_data}
\alias{get_growth_data}
\title{Get dataframe of resampled growth data}
\usage{
get_growth_data(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A \code{qsip_data} object that has been run through \code{run_growth_calculations()}}
}
\value{
A dataframe of the resampled data
}
\description{
This function will return a dataframe of the resampled growth data embedded in a \code{qsip_data}
object.
}
