% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{get_N_total_it}
\alias{get_N_total_it}
\title{Calculate total abundances at timepoint t}
\usage{
get_N_total_it(qsip_data_object, timepoint = "timepoint", t = 0)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{timepoint}{(\emph{character}) The name of the timepoint column in the source data}

\item{t}{(\emph{numeric}) The value of the timepoint column to filter on}
}
\value{
(\emph{data.frame}) A data frame with feature_id and total abundance at time zero
}
\description{
This function takes a \code{qsip_data} object and calculates the total abundance of
each feature at time zero. This should be done on an early \code{qsip_data} object
that still has time zero data.
}
\details{
Sometimes, a feature will have abundance at a later time point, but no values
for time zero. If a feature has zero abundance at time zero, a warning will
be issued, but the feature will still be included in the output with a starting
abundance of zero.
}
