% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{get_overlap_sizes}
\alias{get_overlap_sizes}
\title{Return the count of features present between different groups}
\usage{
get_overlap_sizes(q)
}
\arguments{
\item{q}{a qSIP2 list object}
}
\value{
a tibble
}
\description{
Return the count of features present between different groups
}
