% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resample_data.R
\name{get_resample_data}
\alias{get_resample_data}
\title{Get dataframe of resampled data}
\usage{
get_resample_data(qsip_data_object, type = "all", pivot = FALSE)
}
\arguments{
\item{qsip_data_object}{A \code{qsip_data} object that has been run through \code{run_resampling()}}

\item{type}{(\emph{string}) The type of data to return: "all", "unlabeled", or "labeled"}

\item{pivot}{(\emph{boolean}) Whether to pivot the data into a long format or keep as wide}
}
\value{
A dataframe of the resampled data
}
\description{
This function will return a dataframe of the resampled data embedded in a \code{qsip_data}
object.
}
