% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_sample_object.R
\name{get_sample_counts}
\alias{get_sample_counts}
\title{Get sample counts from qSIP sample data}
\usage{
get_sample_counts(x, ...)
}
\arguments{
\item{x}{An object of \code{qsip_sample_object} class}
}
\description{
Get sample counts from qSIP sample data
}
