% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Get sample dataframe from qSIP2 object}
\usage{
get_sample_data(sample_data, original_headers = FALSE)
}
\arguments{
\item{sample_data}{(\emph{qsip_sample_data or qsip_data}) A qSIP2 object with sample data}

\item{original_headers}{(\emph{boolean, default: FALSE}) If TRUE, the original column names will be used}
}
\value{
A tibble with the sample data
}
\description{
This function will return the sample data stored in a qSIP2 object as a tibble.
}
