% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_data.R
\name{get_source_data}
\alias{get_source_data}
\title{Get source dataframe from qSIP2 object}
\usage{
get_source_data(source_data, original_headers = FALSE)
}
\arguments{
\item{source_data}{(\emph{qsip_source_data or qsip_data}) A qSIP2 object with source data}

\item{original_headers}{(\emph{boolean, default: FALSE}) If TRUE, the original column names will be used}
}
\value{
A tibble with the source data
}
\description{
This function will return the source data stored in a qSIP2 object as a tibble.
}
