% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_source_mat_ids}
\alias{get_source_mat_ids}
\title{Return the source_mat_ids in a qsip object}
\usage{
get_source_mat_ids(qsip_data_object, filtered = FALSE)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} object that has been run through \code{run_resampling()}}

\item{filtered}{(\emph{Boolean}) If TRUE, return the feature_ids from the filtered data}
}
\description{
Return the source_mat_ids in a qsip object
}
