% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_position_validation.R
\name{gradient_position_validation}
\alias{gradient_position_validation}
\title{Check the validity of gradient position values}
\usage{
gradient_position_validation(gradient_position)
}
\arguments{
\item{gradient_position}{A string or vector of gradient position values}
}
\value{
Returns \code{NULL} if the gradient position values are valid, or a printed error
}
\description{
Valid gradient positions are integers. The value can be \code{-1} to represent "bulk" or non-qSIP samples.
}
\keyword{validation}
