% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{ig_get_df_for_resampling}
\alias{ig_get_df_for_resampling}
\title{Make nested dataframe for gpd adjustment sampling}
\usage{
ig_get_df_for_resampling(
  qsip_data_object,
  wad_reference,
  fraction_cutoff,
  source_cutoff
)
}
\description{
Make nested dataframe for gpd adjustment sampling
}
\keyword{internal}
