% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{infer_source_data}
\alias{infer_source_data}
\title{Generate a source data frame from a sample data frame}
\usage{
infer_source_data(sample_data, source_mat_id)
}
\arguments{
\item{sample_data}{(\emph{dataframe}) A data frame with combined sample and source data}

\item{source_mat_id}{(\emph{string}) The column with source_mat_id}
}
\value{
A data frame with the inferred source data
}
\description{
There may be situations where you have a verbose sample data frame that has all
of the data for both the samples and the source materials. This function attempts
to infer a source data frame from the sample data frame. It does this by grouping
by the source matrix ID and then checking to see if all of the values in each
column are the same.If they are, then that column is kept in the source data frame.
If not, then that column is dropped.
}
\seealso{
Other "Sample Data":
\code{\link[=add_gradient_pos_rel_amt]{add_gradient_pos_rel_amt()}}
}
\concept{"Sample Data"}
