% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{infer_source_feature_table}
\alias{infer_source_feature_table}
\title{Infer source feature table from sample counts}
\usage{
infer_source_feature_table(qsip_data_object, copy_number_col)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{copy_number_col}{(\emph{string}) column in the original source metadata with feature copy number}
}
\value{
a feature tibble
}
\description{
This function will convert the tube_relative_abundance data for each feature
into a copy number that was measured for the source material. The resulting
table will have values in the original units of the \code{copy_number_col} parameter.
In other words, if the \code{copy_number_col} is "16S copies per g/soil", then the
output will be in copies of each feature per g/soil.
}
