% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qsip.R
\name{is_multi_qsip_data}
\alias{is_multi_qsip_data}
\title{Validate a multi-qsip list object}
\usage{
is_multi_qsip_data(object, error = FALSE)
}
\arguments{
\item{qsip_list}{A list of qsip_data objects}
}
\description{
Validate a multi-qsip list object
}
