% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qsip.R
\name{is_qsip_data}
\alias{is_qsip_data}
\title{Check object is qsip_data type}
\usage{
is_qsip_data(object, error = FALSE)
}
\arguments{
\item{object}{The object to check if it is a qsip_data object}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Check object is qsip_data type
}
