% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qsip.R
\name{is_qsip_data_list}
\alias{is_qsip_data_list}
\title{Validate a multi-qsip list object}
\usage{
is_qsip_data_list(object, error = FALSE)
}
\arguments{
\item{object}{A list of qsip_data objects}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Validate a multi-qsip list object
}
