% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_qsip_EAF}
\alias{is_qsip_EAF}
\title{Validate a qsip object has EAF values}
\usage{
is_qsip_EAF(object, error = FALSE)
}
\arguments{
\item{object}{The object to check if it is a qsip_data object with EAF calculations}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Validate a qsip object has EAF values
}
