% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_qsip_filtered}
\alias{is_qsip_filtered}
\title{Validate a qsip object has been filtered}
\usage{
is_qsip_filtered(object, error = FALSE)
}
\arguments{
\item{object}{The object to check if it is a filtered qsip_data object}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Validate a qsip object has been filtered
}
