% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qsip.R
\name{is_qsip_growth}
\alias{is_qsip_growth}
\title{Validate a qsip object has been run through growth workflow}
\usage{
is_qsip_growth(object, error = FALSE)
}
\arguments{
\item{object}{The object to check if it is a growth qsip_data object}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Validate a qsip object has been run through growth workflow
}
