% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_qsip_resampled}
\alias{is_qsip_resampled}
\title{Validate a qsip object has been resampled}
\usage{
is_qsip_resampled(object, error = FALSE)
}
\arguments{
\item{object}{The object to check if it is a resampled qsip_data object}

\item{error}{If TRUE it stops with an error message. If FALSE it doesn't error, but returns FALSE}
}
\description{
Validate a qsip object has been resampled
}
