% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotope_palette}
\alias{isotope_palette}
\title{Color palettes for isotopes}
\format{
\subsection{\code{isotope_palette}}{

A named vector with length 6
}
}
\usage{
isotope_palette
}
\description{
Used by default in the plots
}
\seealso{
Other "palettes":
\code{\link{source_palette}}
}
\concept{"palettes"}
\keyword{datasets}
