% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_validation.R
\name{isotope_validation}
\alias{isotope_validation}
\title{Check the validity of an isotope string}
\usage{
isotope_validation(
  isotope,
  isotope_list = c("12C", "13C", "14N", "15N", "16O", "18O")
)
}
\arguments{
\item{isotope}{Isotope value or values (string(s))}

\item{isotope_list}{Isotopes to check against (strings, default: c("12C", "13C", "14N", "15N", "16O", "18O"))}
}
\value{
Returns \code{NULL} if the isotope strings are valid, or a printed error
}
\description{
Check the validity of an isotope string
}
\note{
The isotope_list may change if isotopolog_label stays a thing. Only the "heavy" isotopes will be allowed.
}
\keyword{internal}
