% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jgi_mixes}
\alias{jgi_mixes}
\title{JGI control sequences}
\format{
\subsection{\code{jgi_mixes}}{

252x8 tibble
}
}
\usage{
jgi_mixes
}
\description{
Used to normalize controls
}
\concept{"normalization"}
\keyword{datasets}
