% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{jgi_sample_df}
\alias{jgi_sample_df}
\title{Make a sample dataframe from JGI proposal file}
\usage{
jgi_sample_df(file_path, sources, skip = 27)
}
\arguments{
\item{file_path}{A file path to a JGI proposal file}

\item{sources}{A dataframe with source information}

\item{skip}{The number of lines to skip in the proposal file}
}
\value{
A dataframe suitable for \code{qSIP2} workflows
}
\description{
Make a sample dataframe from JGI proposal file
}
\seealso{
Other "spike-ins":
\code{\link[=fit_regression_model]{fit_regression_model()}},
\code{\link[=get_normalized_controls]{get_normalized_controls()}},
\code{\link[=get_normalized_features]{get_normalized_features()}},
\code{\link[=jgi_feature_df]{jgi_feature_df()}},
\code{\link[=jgi_normalize_features]{jgi_normalize_features()}},
\code{\link[=jgi_source_df]{jgi_source_df()}}
}
\concept{"spike-ins"}
