% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{make_delta_EAF_contrasts}
\alias{make_delta_EAF_contrasts}
\title{Make an all-by-all contrasts table for delta EAF calculations}
\usage{
make_delta_EAF_contrasts(q)
}
\arguments{
\item{q}{a qSIP2 list object}
}
\value{
a dataframe
}
\description{
This is an internal function that runs if a user does not provide a contrasts
table to \code{run_delta_EAF_contrasts()}. It will produce an all-by-all table
and run through \code{validate_delta_EAF_contrasts()} to modify the class.
}
