% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_wrapper.R
\name{multi_qsip_wrapper_launcher}
\alias{multi_qsip_wrapper_launcher}
\title{Internal function for launching multi_qsip_wrapper()}
\usage{
multi_qsip_wrapper_launcher(
  group,
  name,
  qsip_data_object,
  min_unlabeled_sources = 2,
  min_labeled_sources = 2,
  min_unlabeled_fractions = 2,
  min_labeled_fractions = 2,
  allow_failures = FALSE,
  resamples = 1000,
  seed = NULL
)
}
\description{
Internal function for launching multi_qsip_wrapper()
}
