% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_resamples.R
\name{n_resamples}
\alias{n_resamples}
\title{Number of resamples}
\usage{
n_resamples(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} object that has been run through \code{run_resampling()}}
}
\description{
Number of resamples
}
