% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbase.R
\name{pivot_kbase_amplicon_matrix}
\alias{pivot_kbase_amplicon_matrix}
\title{Pivot KBase amplicon matrix}
\usage{
pivot_kbase_amplicon_matrix(amplicon)
}
\arguments{
\item{amplicon}{An amplicon matrix from KBase}
}
\description{
Pivot KBase amplicon matrix
}
\keyword{internal}
