% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{plot_corrected_curves}
\alias{plot_corrected_curves}
\title{Plot curves of corrected sample curves versus reference}
\usage{
plot_corrected_curves(reference, corrected)
}
\arguments{
\item{reference}{(\emph{qsip_data}) A qsip data object}

\item{corrected}{(\emph{qsip_data}) A qsip data object}
}
\description{
Plot curves of corrected sample curves versus reference
}
