% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{plot_correction_compression}
\alias{plot_correction_compression}
\title{Plot correction compression}
\usage{
plot_correction_compression(corrections)
}
\arguments{
\item{corrections}{(\emph{tibble}) A tibble of corrections from \code{correct_gpd_rlm()}}
}
\description{
Plot correction compression
}
