% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{plot_correction_stats}
\alias{plot_correction_stats}
\title{Plot the correction statistics}
\usage{
plot_correction_stats(
  reference,
  corrected,
  source_cutoff = 3,
  fraction_cutoff = 5,
  quiet = F
)
}
\arguments{
\item{reference}{(\emph{qsip_data}) A qsip data object}

\item{corrected}{(\emph{qsip_data}) A qsip data object}

\item{source_cutoff}{(\emph{numeric, default: 3}) The minimum number of sources a feature must be found in to be considered}

\item{fraction_cutoff}{(\emph{numeric, default: 5}) The minimum number of fractions a feature must be found in to be considered}

\item{quiet}{(\emph{logical, default: FALSE}) Whether to print a message about the number of features found}
}
\description{
Plot the correction statistics
}
