% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density_outliers}
\alias{plot_density_outliers}
\title{Cook's outlier detection on gradient positions vs densities}
\usage{
plot_density_outliers(sample_data, sensitivity = 4)
}
\arguments{
\item{sample_data}{(\emph{qsip_sample_data or qsip_data}) A qsip object with sample data}

\item{sensitivity}{(\emph{numeric, default: 4}) A sensitivity value, with lower values being more sensitive to outlier detection}
}
\value{
A ggplot object
}
\description{
Assuming a linear relationship between the gradient_position and gradient_pos_density,
this function will plot any potential outliers using Cook's distance and a defined
sensitivity. The lower the sensitivity, the more likely outliers will be flagged.
}
\seealso{
Other "visualizations":
\code{\link[=plot_feature_curves]{plot_feature_curves()}},
\code{\link[=plot_filter_results]{plot_filter_results()}},
\code{\link[=plot_sample_curves]{plot_sample_curves()}},
\code{\link[=plot_source_wads]{plot_source_wads()}}
}
\concept{"visualizations"}
