% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{plot_difference_to_mean}
\alias{plot_difference_to_mean}
\title{Plot difference between feature WAD and WAD reference}
\usage{
plot_difference_to_mean(
  qsip_data_object,
  fraction_cutoff = 5,
  source_cutoff = 3
)
}
\description{
Plot difference between feature WAD and WAD reference
}
