% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_EAF_values}
\alias{plot_EAF_values}
\title{Plot EAF and confidence intervals}
\usage{
plot_EAF_values(
  qsip_data_object,
  confidence = 0.9,
  success_ratio = 0.9,
  top = Inf,
  error = "none",
  alpha = 0.3,
  zero_line = TRUE,
  shared_y = FALSE,
  title = NULL
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip_data object or list of qsip_data objects}

\item{confidence}{(\emph{numeric}) The confidence level for the confidence interval}

\item{success_ratio}{(\emph{numeric}) The ratio of successful resamples to total resamples}

\item{top}{(\emph{numeric}) The number of top features to plot. Use \code{Inf} for all}

\item{error}{(\emph{character}) The type of error bars to plot. Options are 'none', 'bar', 'ribbon'}

\item{alpha}{(\emph{numeric}) The transparency of the error bar/ribbon}

\item{zero_line}{(\emph{logical}) Add a line at EAF = 0}

\item{shared_y}{(\emph{logical}) Use a shared y-axis for the facets}

\item{title}{(\emph{character}) An optional title of the plot}
}
\description{
This function plots the observed EAF values for each feature in the dataset.
The features are ordered by their observed EAF values. The confidence intervals
are plotted as error bars or ribbons.
}
\details{
Either a single qsip object or a list of named qsip objects can be passed to this
function. If giving a list of qsip objects the plot will be faceted by the
list names.

If the resampling step was run with the default \code{allow_failures = FALSE}, then the
points will just be colored a generic blue. But, if resampling was instead run
with \code{allow_failures = TRUE}, then the points are colored based on the success
ratio of the resamples. If giving a list of qsip objects then the pass/fail color scheme
will be applied if any of the qsip objects have \code{allow_failures = TRUE}.
}
