% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_feature_resamplings}
\alias{plot_feature_resamplings}
\title{Plot the resampled EAFs for each feature}
\usage{
plot_feature_resamplings(
  qsip_data_object,
  feature_ids = NULL,
  area = TRUE,
  confidence = 0.9,
  intervals = ""
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object that has been resampled}

\item{feature_ids}{(\emph{character vector}) A vector of feature ids to filter on}

\item{area}{(\emph{boolean}) Whether to plot the area under the curve or not (default: TRUE)}

\item{confidence}{(\emph{numeric}) The confidence interval to plot}

\item{intervals}{(\emph{character}) Whether to plot the confidence interval as a bar, line or not at all (default)}
}
\value{
A ggplot object
}
\description{
This plot will show the results of the resampling procedure for each feature.
The resampling procedure is run using the run_resampling() function. The plot
will show the mean resampled EAF for each feature, with the confidence interval
(default 90\%) shown as a bar or line (default no line). The area under the curve
can also be shown (default TRUE).
}
