% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_filter_results}
\alias{plot_filter_results}
\title{Plot the results of filtering features}
\usage{
plot_filter_results(
  qsip_data_object,
  return_type = "combined",
  colors = lifecycle::deprecated()
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data})}

\item{return_type}{(\emph{string, default: combined}) Changes the return type from a combined plot (\emph{combined}), list of individual plots (\emph{individual}) or list of dataframes (\emph{dataframe})}

\item{colors}{deprecated}
}
\value{
Combined or individual plots, or the raw dataframes
}
\description{
After running the \code{run_feature_filter()} code, this function will produce two
plot detailing the consequences of the filtering. Plot A shows the retained and
removed features by their total tube relative abundance contribution, and plot B
shows the retained and removed features by the total count of each category. The
"zero fractions" shown in plot B are those entirely missing from the given source_mat_id
and are thus not going to be retained no matter what the \code{run_filter_feature()}
parameters are.
}
\seealso{
Other "visualizations":
\code{\link[=plot_density_outliers]{plot_density_outliers()}},
\code{\link[=plot_feature_curves]{plot_feature_curves()}},
\code{\link[=plot_sample_curves]{plot_sample_curves()}},
\code{\link[=plot_source_wads]{plot_source_wads()}}
}
\concept{"visualizations"}
