% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_source_wads}
\alias{plot_source_wads}
\title{Plot the source WADs by isotope}
\usage{
plot_source_wads(qsip_data, group = NULL, title = NULL)
}
\arguments{
\item{qsip_data}{(\emph{qsip_data}) qSIP object}

\item{group}{(\emph{character}) An optional grouping parameter to facet the y or x,y axes.
Can be a single column name or vector of column names from source metadata.
Multiple columns are combined with "_" separator
(e.g., \code{group = c("Moisture", "isotopolog")} creates "Drought_glucose", "Normal_glucose").}

\item{title}{(\emph{character}) An optional title for the plot}
}
\value{
A ggplot object
}
\description{
Plot the source WADs by isotope
}
\seealso{
Other "visualizations":
\code{\link[=plot_density_outliers]{plot_density_outliers()}},
\code{\link[=plot_feature_curves]{plot_feature_curves()}},
\code{\link[=plot_filter_results]{plot_filter_results()}},
\code{\link[=plot_sample_curves]{plot_sample_curves()}}
}
\concept{"visualizations"}
