% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_successful_resamples}
\alias{plot_successful_resamples}
\title{Plot the number of successful resamples for each feature_id}
\usage{
plot_successful_resamples(
  qsip_data_object,
  labels = FALSE,
  as_percentage = FALSE
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object that has been resampled}

\item{labels}{(\emph{logical}) If TRUE, the counts will be labeled on the plot}

\item{as_percentage}{(\emph{logical}) If TRUE, the x axis will be the percentage of the total number of resamples}
}
\value{
A ggplot object
}
\description{
This function will plot the number of successful resamples for each feature_id.
This value will typically be the number of resamples given to \code{run_resampling()},
but if \code{run_resampling()} is called with \code{allow_failures = TRUE} then the number
of successful resamples might be less than the number of resamples given.
}
