% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP2-package.R
\docType{package}
\name{qSIP2-package}
\alias{qSIP2}
\alias{qSIP2-package}
\title{qSIP2: qSIP Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Package for analyzing quantitative stable isotope probing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeffkimbrel/qSIP2}
  \item \url{https://jeffkimbrel.github.io/qSIP2/}
  \item Report bugs at \url{https://github.com/jeffkimbrel/qSIP2/issues}
}

}
\author{
\strong{Maintainer}: Jeff Kimbrel \email{kimbrel1@llnl.gov} (\href{https://orcid.org/0000-0001-7213-9392}{ORCID})

Authors:
\itemize{
  \item Jeff Kimbrel \email{kimbrel1@llnl.gov} (\href{https://orcid.org/0000-0001-7213-9392}{ORCID})
}

}
\keyword{internal}
