% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_abundance_object.R
\name{qsip_abundance_object}
\alias{qsip_abundance_object}
\title{qSIP abundance table class}
\usage{
qsip_abundance_object(
  data = class_missing,
  id = class_missing,
  taxonomy = class_missing
)
}
\description{
A class to hold and validate an abundance table
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe or tibble}

\item{\code{id}}{Column name with unique taxa IDs}
}}

\keyword{object}
