% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_objects.R
\name{qsip_clean}
\alias{qsip_clean}
\title{Clean large data from qSIP object}
\usage{
qsip_clean(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qSIP data object}
}
\description{
Clean large data from qSIP object
}
