% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_experiment_object.R
\name{qsip_experiment_object}
\alias{qsip_experiment_object}
\title{qSIP experiment data class}
\usage{
qsip_experiment_object(data, isotope, isotopolog, source_mat_id)
}
\value{
A \code{qsip_experiment_object} object
}
\description{
A class to hold and validate qSIP experiment data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Experiment metadata}

\item{\code{isotope}}{Column name with the isotope name}

\item{\code{isotopolog}}{Column name with the MISIP isotopolog data}

\item{\code{source_mat_id}}{The unique ID for the biological subject or replicate}

\item{\code{source_dna_ng}}{xxx}
}}

\keyword{object}
