% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_object_size.R
\name{qsip_object_size}
\alias{qsip_object_size}
\title{Return the size of qsip data objects or each slot}
\usage{
qsip_object_size(qsip_data_object, units = "auto")
}
\arguments{
\item{qsip_data_object}{A qsip object or list of objects}

\item{units}{Units to format the size in. Default is "auto"}
}
\description{
Return the size of qsip data objects or each slot
}
