% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_taxa_object.R
\name{qsip_taxa_object}
\alias{qsip_taxa_object}
\title{qSIP taxa table class}
\usage{
qsip_taxa_object(data = class_missing, id = class_missing)
}
\description{
A class to hold and validate taxa abundance table
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe or tibble}

\item{\code{id}}{Column name with unique taxa IDs}
}}

\keyword{sample_data}
