% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{run_delta_EAF_contrasts}
\alias{run_delta_EAF_contrasts}
\title{Get delta EAF values between contrast groupings}
\usage{
run_delta_EAF_contrasts(q, contrasts = NULL, confidence = 0.95, quiet = FALSE)
}
\arguments{
\item{q}{a qSIP2 list object}

\item{contrasts}{A validated contrasts table}

\item{confidence}{(numeric, default: 0.95) confidence interval}
}
\value{
a tibble
}
\description{
Given a list of qSIP2 objects, this function will return the delta EAF value
for each feature_id, including the confidence interval, standard deviation,
and two p-values from 1) the bootstrap percentile test (\code{bs_pval}) and 2) the
Wald's z-test (\code{pval}).
}
\details{
By default the function will do all pairwise comparisons of the groups in the
object list with the first listed as the "control" and the second as the
"treatment". Optionally, data frame with "control", "treatment" and
"contrast" columns can be provided to specify only the contrasts of interest
and to define the proper "control" vs. "treatment" groups. To ensure this
data frame is properly formatted (e.g. has the proper column names and all
groups are found in the qSIP2 object list), it must be validated through the
\code{validate_delta_EAF_contrasts()} function.

Under the hood, for each feature_id this function subtracts the observed EAF
in the "treatment" group from the "control" and saves this as the \eqn{\delta}
value. Then, it pulls the data from \code{run_resampling()} and for each resampling
\eqn{i} calculates the delta (so \eqn{\delta_i = control_i - treatment_i}) saving
these values as a vector of delta distributions. The \code{bs_pval} is calculated
from this distribution, and \code{pval} is calculated from \eqn{\delta} and the
standard deviation of the distribution.
}
