% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_feature_filter.R
\name{run_feature_filter}
\alias{run_feature_filter}
\title{Filter features in a qSIP data object}
\usage{
run_feature_filter(
  qsip_data_object,
  unlabeled_source_mat_ids,
  labeled_source_mat_ids,
  min_unlabeled_sources = 2,
  min_labeled_sources = 2,
  min_unlabeled_fractions = 2,
  min_labeled_fractions = 2,
  quiet = FALSE
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{unlabeled_source_mat_ids}{(\emph{string or strings(s)}) A list of the unlabeled source_mat_ids to filter on}

\item{labeled_source_mat_ids}{(\emph{string or strings(s)}) A list of the labeled source_mat_ids to filter on}

\item{min_unlabeled_sources}{(\emph{integer, default: 2}) Minimum number of unlabeled source_mat_ids a feature must be found in.}

\item{min_labeled_sources}{(\emph{integer, default: 2}) Minimum number of labeled source_mat_ids a feature must be found in.}

\item{min_unlabeled_fractions}{(\emph{integer, default: 2}) Minimum number of fractions a feature must be found in to be present in an unlabeled source_mat_id}

\item{min_labeled_fractions}{(\emph{integer, default: 2}) Minimum number of fractions a feature must be found in to be present in an labeled source_mat_id}
}
\value{
An updated \code{qsip_data} object with a filtered feature dataframe in the
\verb{@filtered_feature_data} slot and intermediate data in the \verb{@filter_results}
slot for plotting.
}
\description{
Filters a feature in a feature table in a \code{qsip_data} object by presence in
a minimum number of fractions and source_mat_ids.
}
\details{
Filtering is first done on the fractions, and then on the source_mat_ids. For
example, take a feature that is found in three source_mat_ids in 3, 5 and 9
fractions. If you set \code{min_fractions = 5} and \code{min_sources = 3} then this
feature will not survive the filtering because although it is found in three
source_mat_ids, one of them is less then the minimum fraction count and would
therefore be considered not found in that source_mat_id.

The feature table is filtered to retain only the feature_ids passing the
filter, and to keep only the sample_ids that correspond to the given
source_mat_ids. This filtered table is stored in the \verb{@filtered_feature_data}
slot, and the values in this table are not the raw initial values but are the
relative abundances per fraction per tube. These values are originally
calculated during \code{qsip_data} object creation and the values for all features
are stored in the \verb{@tube_rel_abundance} slot.
}
\concept{"qSIP Filtering"}
