% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{run_growth_calculations}
\alias{run_growth_calculations}
\title{Run growth calculations}
\usage{
run_growth_calculations(qsip_data_object, N_total_it, timepoint = "timepoint")
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{N_total_it}{(\emph{data.frame}) A data frame of time zero totals from \code{get_N_total_it()}}
}
\description{
Run growth calculations
}
