% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_resampling.R
\name{run_resampling}
\alias{run_resampling}
\title{Resample WAD values}
\usage{
run_resampling(
  qsip_data_object,
  resamples = 1000,
  with_seed = NULL,
  progress = TRUE
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object that has been filtered}

\item{resamples}{(\emph{integer}) The number of resamples/bootstraps to run}

\item{with_seed}{(\emph{integer}) An optional seed for reproducibility}

\item{progress}{(\emph{boolean}) Option to show a progress bar for the resampling step}
}
\value{
A new \code{qsip_data} object with the \verb{@resamples} slot populated with resamples wad values
}
\description{
Takes a filtered WAD dataframe and resamples x times
}
\details{
This function returns a list of resampled dataframes of x length for both the
labeled and unlabeled sources.
}
