% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_nonMISIP}
\alias{sample_data_nonMISIP}
\title{Example qSIP Sample Data (non MISIP)}
\format{
\subsection{\code{sample_data_nonMISIP}}{

A data frame with 72 rows and 5 columns:
\describe{
\item{rep}{The unique name of the biological replicate}
\item{isotope}{The heavy isotope}
\item{fraction}{The gradient position ranked from highest to lowest density}
\item{density}{The density of the fraction}
\item{dna_conc}{The DNA concentration of the fraction}
}
}
}
\usage{
sample_data_nonMISIP
}
\description{
An example sample data tibble that is not formatted according to MISIP standards.
}
\keyword{data}
