% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{show_unshared_ids}
\alias{show_unshared_ids}
\title{Show missing source_mat_ids and sample_ids (deprecated)}
\usage{
show_unshared_ids(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qSIP data object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
