% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source_palette}
\alias{source_palette}
\title{An expanding color palette}
\format{
\subsection{\code{source_palette}}{

A function that takes a number
}
}
\usage{
source_palette
}
\description{
Used in some plots
}
\seealso{
Other "palettes":
\code{\link{isotope_palette}}
}
\concept{"palettes"}
\keyword{datasets}
