% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_EAF_values.R
\name{summarize_EAF_values}
\alias{summarize_EAF_values}
\title{Summarize the observed and resampled EAF values}
\usage{
summarize_EAF_values(qsip_data_object, confidence = 0.9)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}

\item{confidence}{(\emph{numeric, default: 0.9}) The confidence level for the confidence interval}
}
\value{
A \code{dataframe} with summarized observed and resampled EAF values
}
\description{
Reports observed EAF value (\code{observed_EAF}) as well as the mean of the resampled
values (\code{mean_resampled_EAF}) and the \code{lower} and \code{upper} confidence interval
with a given \code{confidence} limit.
}
\details{
The confidence interval uses the resampling method where it returns the quantile
values from the resampled data. If \code{confidence = 0.9} (the default) then this
function returns the 5\% and 95\% quantiles (representing 90\% of the resampling)
as the \code{lower} and \code{upper} results.
}
