% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_EAF_values.R
\name{summarize_EAF_values_internal}
\alias{summarize_EAF_values_internal}
\title{Internal function to summarize EAF values}
\usage{
summarize_EAF_values_internal(qsip_data_object, confidence = 0.9)
}
\description{
Called by \code{summarize_EAF_values} to calculate the resampled EAF values.
}
