% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{summarize_growth_values}
\alias{summarize_growth_values}
\title{Summarize growth values}
\usage{
summarize_growth_values(qsip_data_object, confidence = 0.9, quiet = FALSE)
}
\description{
Summarize growth values
}
