% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_isotope_names}
\alias{valid_isotope_names}
\title{Valid isotope names}
\format{
\subsection{\code{valid_isotope_names}}{

a list of length 3
}
}
\usage{
valid_isotope_names
}
\description{
Used to make sure isotope names are valid
}
\keyword{datasets}
