% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{validate_delta_EAF_contrasts}
\alias{validate_delta_EAF_contrasts}
\title{Validate a delta EAF contrasts table}
\usage{
validate_delta_EAF_contrasts(q, contrasts)
}
\arguments{
\item{q}{a qSIP2 list object}

\item{contrasts}{A validated contrasts table}
}
\value{
a validated contrasts table
}
\description{
Validate a delta EAF contrasts table
}
