% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gradient_position.R
\name{validate_gradient_position}
\alias{validate_gradient_position}
\title{Check the validity of gradient position values (internal)}
\usage{
validate_gradient_position(gradient_position)
}
\arguments{
\item{gradient_position}{(\emph{string or strings}) Gradient position value or values}
}
\value{
\code{NULL} if the gradient position values are valid, or a
printed error
}
\description{
Valid gradient positions are integers. The value can be \code{-1} to represent
"bulk" or non-qSIP samples.
}
