% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_source_isotope.R
\name{validate_source_isotope}
\alias{validate_source_isotope}
\title{Validate the given source mat ids have the expected labeled/unlabeled designation (internal)}
\usage{
validate_source_isotope(qsip_data_object, source_mat_ids, isotope_list)
}
\value{
TRUE (all match) or FALSE (some don't match)
}
\description{
Currently used by run_feature_filter() to make sure user given labeled or unlabeled
source_mat_ids are not incorrect with respective to their source_data.
}
