% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_abundance_validation.R
\name{abundance_validation}
\alias{abundance_validation}
\title{Check the validity of an abundance table}
\usage{
abundance_validation(data, id)
}
\arguments{
\item{data}{ASV/OTU table}

\item{id}{column name with taxa IDs}
}
\value{
Returns \code{NULL} if the values are valid, or a printed error
}
\description{
This validation function will check the validity of an abundance table.
}
\details{
Rows should contain the unique taxa ids with a column designated with the id
argument. Each other column should be a unique sample name.

Validity checking includes making sure all data is numeric (except for the
taxa IDs), all numbers are integers, and no numbers are negative.
}
\keyword{validation}
