% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{add_gradient_pos_rel_amt}
\alias{add_gradient_pos_rel_amt}
\title{Add gradient_pos_rel_amt to data}
\usage{
add_gradient_pos_rel_amt(
  data,
  amt,
  source_mat_id = "source_mat_id",
  overwrite = FALSE
)
}
\arguments{
\item{data}{(\emph{dataframe}) Sample metadata}

\item{amt}{(\emph{string}) Column name that has the qPCR or DNA amounts per fraction}

\item{source_mat_id}{(\emph{string, default: "source_mat_id"}) Column name with the source_mat_id}

\item{overwrite}{(\emph{bool, default: FALSE}) Determines whether or not to overwrite an existing gradient_pos_rel_amt column}
}
\value{
A dataframe with a \code{gradient_pos_rel_amt} column
}
\description{
This function will calculate the relative amount of a fraction compared to
the whole source using either qPCR copies or DNA concentrations. This works
by summing all of the \if{html}{\out{<amt>}} within a source, and then dividing each fraction
by the total. The total percent will always equal 1 (or 100\%) when running
this function.
}
\details{
There may be times that you would want to provide values that do not sum to
100\%. For example, if you throw out a certain fraction because it didn't
sequence well, then your total could be less than 100\%.
}
\seealso{
Other "Sample Data": 
\code{\link{infer_source_data}()}
}
\concept{"Sample Data"}
