% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gradient_pos_rel_amt.R
\name{add_gradient_pos_rel_amt}
\alias{add_gradient_pos_rel_amt}
\title{Add gradient_pos_rel_amt to data}
\usage{
add_gradient_pos_rel_amt(
  data,
  source_mat_id = "source_mat_id",
  amt,
  overwrite = F
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{source_mat_id}{Grouping variable for a replicate}

\item{amt}{Column name that has the qPCR or DNA amounts per fraction}

\item{overwrite}{Determines whether or not to overwrite an existing gradient_pos_rel_amt column}
}
\description{
This function will calculate the relative amount of a fraction compared to
the whole replicate using either qPCR copies or DNA concentrations.
}
\keyword{sample_data}
