% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gradient_pos_rel_amt.R
\name{add_gradient_pos_rel_amt}
\alias{add_gradient_pos_rel_amt}
\title{Add gradient_pos_rel_amt to data}
\usage{
add_gradient_pos_rel_amt(
  data,
  amt,
  source_mat_id = "source_mat_id",
  overwrite = FALSE
)
}
\arguments{
\item{data}{(\emph{dataframe}) Sample metadata}

\item{amt}{(\emph{string}) Column name that has the qPCR or DNA amounts per fraction}

\item{source_mat_id}{(\emph{string, default: "source_mat_id"}) Grouping variable for a replicate}

\item{overwrite}{(\emph{bool, default: FALSE}) Determines whether or not to overwrite an existing gradient_pos_rel_amt column}
}
\value{
A dataframe with a \code{gradient_pos_rel_amt} column
}
\description{
This function will calculate the relative amount of a fraction compared to
the whole replicate using either qPCR copies or DNA concentrations.
}
\concept{"Sample Data"}
