% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_isotopolog_label.R
\name{add_isotopolog_label}
\alias{add_isotopolog_label}
\title{Add isotopolog_label to source data}
\usage{
add_isotopolog_label(data, isotope = "isotope")
}
\arguments{
\item{data}{(\emph{dataframe}) Sample metadata}

\item{isotope}{(\emph{string, default: "isotope"}) Column name with isotope data}
}
\value{
A dataframe with \code{isotopolog_label} column added and \code{isotope} column
modified
}
\description{
To convert from an isotope only dataframe to one with both isotope and
isotopolog_label to satisfy MISIP requirements
}
\concept{"Source Data"}
