% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_isotopolog_label.R
\name{add_isotopolog_label}
\alias{add_isotopolog_label}
\title{Add isotopolog_label to data}
\usage{
add_isotopolog_label(data, isotope)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{isotope}{Column name isotope data}
}
\description{
To convert from an isotope only dataframe to one with both isotope and isotopolog_label to satisfy MISIP requirements
}
\details{
Note, this function also renames your isotope column to "isotope", if it isn't already.
}
\keyword{sample_data}
