% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxonomy.R
\name{add_taxonomy}
\alias{add_taxonomy}
\title{Add a taxonomy table to qSIP abundance data}
\usage{
add_taxonomy(feature_object, taxa, feature_id)
}
\arguments{
\item{feature_object}{(\emph{qsip_feature_data}) An object of \code{qsip_feature_data} class}

\item{taxa}{(\emph{dataframe}) A taxa table}

\item{id}{(\emph{string}) The column name for the taxa ids that match the ids in the
abundance table}
}
\value{
An updated \code{qsip_feature_data} with the taxonomy slot populated with a taxonomy dataframe.
}
\description{
Add a taxonomy table to qSIP abundance data
}
\concept{"Feature Data"}
