% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_atoms}
\alias{calculate_atoms}
\title{Get the number of relevant atoms per nucleotide (internal)}
\usage{
calculate_atoms(G, isotope)
}
\arguments{
\item{G}{(\emph{numeric}) GC percentage}

\item{isotope}{(\emph{string}) The isotope to use for calculations... either 13C, 15N or 18O}
}
\value{
(\emph{numeric}) The number of atoms per nucleotide
}
\description{
Carbon and nitrogen atoms varies with GC content, but oxygen content is
constant for DNA regardless of GC content.
}
\keyword{internal}
