% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{calculate_bi}
\alias{calculate_bi}
\title{Calculate birth rate}
\usage{
calculate_bi(N_total_it, N_light_it, timepoint, timepoint1)
}
\arguments{
\item{N_total_it}{The copy number of feature i at timepoint t}

\item{N_light_it}{The copy number of feature i at timepoint 0 (or timepoint that is being compared against)}

\item{timepoint}{The timepoint at which the copy number is being measured}

\item{timepoint1}{The timepoint that is being compared against}
}
\description{
Equation 7 from Koch, 2018
}
