% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_ce}
\alias{calculate_ce}
\title{Calculate Cumulative Enrichment (CE) from qSIP object(s)}
\usage{
calculate_ce(
  qsip_data_object,
  confidence = 0.95,
  isotope_label = c("labeled", "unlabeled", "both")
)
}
\arguments{
\item{qsip_data_object}{A single qSIP object or list of qSIP objects}

\item{confidence}{Confidence level for EAF intervals (default 0.95)}

\item{isotope_label}{Which samples to calculate CE for: "labeled", "unlabeled", or "both" (default "labeled")}
}
\value{
Data frame with CE calculations per sample
}
\description{
Calculate Cumulative Enrichment (CE) from qSIP object(s)
}
