% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{completely_labeled_values}
\alias{completely_labeled_values}
\title{Returns the max labeling of a given isotope (internal)}
\usage{
completely_labeled_values(isotope)
}
\arguments{
\item{isotope}{The heavy isotope (13C, 15N or 18O)}
}
\value{
A number representing the max labeling of a given isotope
}
\description{
Returns the max labeling of a given isotope (internal)
}
