% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{correct_gpd_bootstrap}
\alias{correct_gpd_bootstrap}
\title{Correct gradient position density (gpd) values using bootstrapped features}
\usage{
correct_gpd_bootstrap(
  qsip_data_object,
  boostraps = 1000,
  fraction_cutoff = 5,
  source_cutoff = 3
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}
}
\description{
Correct gradient position density (gpd) values using bootstrapped features
}
