% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{correct_gpd_rlm}
\alias{correct_gpd_rlm}
\title{Correct gradient position density (gpd) values using linear regression approach}
\usage{
correct_gpd_rlm(
  qsip_data_object,
  method = "siegel",
  fraction_cutoff = 5,
  source_cutoff = 3,
  quiet = F,
  return = "qsip_data_object"
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}

\item{method}{(\emph{character}) The lm method to use}

\item{fraction_cutoff}{(\emph{numeric, default: 5}) The minimum number of fractions a feature must be found in to be considered}

\item{source_cutoff}{(\emph{numeric, default: 3}) The minimum number of sources a feature must be found in to be considered}

\item{quiet}{(\emph{logical, default: FALSE}) Whether to print a message about the number of features found}

\item{return}{(*character, default: "qsip_data_object") Whether to return the corrections or the qsip_data_object}
}
\description{
Correct gradient position density (gpd) values using linear regression approach
}
