% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sample_df}
\alias{example_sample_df}
\title{Example Sample Dataframe}
\format{
\subsection{\code{example_sample_df}}{

A data frame with 284 rows and 6 columns:
\describe{
\item{sample}{The unique sample id of the fraction}
\item{source}{The parent source id for this sample}
\item{Fraction}{The gradient position}
\item{density_g_ml}{The density of this fraction}
\item{dna_conc}{The concentration of DNA post-fractionation}
\item{avg_16S_g_soil}{qPCR amplicon count}
}
}
}
\usage{
example_sample_df
}
\description{
An example sample data table with fraction-related metadata
}
\seealso{
Other "example datasets": 
\code{\link{example_feature_df}},
\code{\link{example_feature_object}},
\code{\link{example_group_dataframe}},
\code{\link{example_qsip_growth_object}},
\code{\link{example_qsip_growth_t0}},
\code{\link{example_qsip_object}},
\code{\link{example_sample_object}},
\code{\link{example_source_df}},
\code{\link{example_source_object}}
}
\concept{"example datasets"}
\keyword{datasets}
