% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shared_ids.R
\name{find_shared_ids}
\alias{find_shared_ids}
\title{Find shared sources and samples in qSIP objects}
\usage{
find_shared_ids(source_data, sample_data, feature_data)
}
\arguments{
\item{source_data}{qsip source_data object}

\item{sample_data}{qsip sample_data object}

\item{feature_data}{qsip feature_data object}
}
\description{
Find shared sources and samples in qSIP objects
}
