% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression_model.R
\name{fit_regression_model}
\alias{fit_regression_model}
\title{Fit regression model to spike-in control data}
\usage{
fit_regression_model(control_data, engine = "lm")
}
\description{
Fit regression model to spike-in control data
}
