% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeins.R
\name{fit_regression_model}
\alias{fit_regression_model}
\title{Fit regression model to spike-in control data}
\usage{
fit_regression_model(control_data, engine = "lm")
}
\arguments{
\item{control_data}{A data frame containing the spike-in control data}

\item{engine}{The regression engine to use. Choices are "lm" or "glm"}
}
\description{
Fit regression model to spike-in control data
}
