% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_objects.R
\name{get_dataframe}
\alias{get_dataframe}
\title{Return the original dataframe from a qsip_feature_data object}
\arguments{
\item{x}{(\emph{qsip_feature_data}) A qSIP feature data object}

\item{original_headers}{(\emph{logical}) Return the original column names}
}
\description{
Return the original dataframe from a qsip_feature_data object
}
