% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_features.R
\name{get_normalized_features}
\alias{get_normalized_features}
\title{Get normalized features from jgi_normalize_features}
\usage{
get_normalized_features(normalized_df)
}
\description{
Get normalized features from jgi_normalize_features
}
