% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_normalized_features}
\alias{get_normalized_features}
\title{Get normalized features from jgi_normalize_features}
\usage{
get_normalized_features(normalized_df)
}
\arguments{
\item{normalized_df}{A dataframe from jgi_normalize_features}
}
\value{
A dataframe suitable for \code{qSIP2} workflows
}
\description{
Get normalized features from jgi_normalize_features
}
\seealso{
Other "spike-ins": 
\code{\link{fit_regression_model}()},
\code{\link{get_normalized_controls}()},
\code{\link{jgi_feature_df}()},
\code{\link{jgi_normalize_features}()},
\code{\link{jgi_sample_df}()},
\code{\link{jgi_source_df}()}
}
\concept{"spike-ins"}
