% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_pos_density_validation.R
\name{gradient_pos_density_validation}
\alias{gradient_pos_density_validation}
\title{Check the validity of density values}
\usage{
gradient_pos_density_validation(gradient_pos_density, low = 1.55, high = 1.8)
}
\arguments{
\item{gradient_pos_density}{Density value or values (string(s))}

\item{low}{A low limit for valid density values (numeric, default: 1.55)}

\item{high}{A high limit for valid density values (numeric, default: 1.8)}
}
\value{
Returns \code{NULL} if the density values are valid, or a printed error
}
\description{
Check the validity of density values
}
\keyword{validation}
