% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_pos_density_validation.R
\name{gradient_pos_density_validation}
\alias{gradient_pos_density_validation}
\title{Check the validity of a density values}
\usage{
gradient_pos_density_validation(gradient_pos_density, low = 1.55, high = 1.8)
}
\arguments{
\item{gradient_pos_density}{A string or vector of density values}

\item{low}{A low limit for valid density values}

\item{high}{A high limit for valid density values}
}
\value{
Returns \code{NULL} if the density values are valid, or a printed error
}
\description{
Check the validity of a density values
}
