% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{ig_get_df_for_resampling}
\alias{ig_get_df_for_resampling}
\title{Make nested dataframe for gpd adjustment sampling}
\usage{
ig_get_df_for_resampling(
  qsip_data_object,
  wad_reference,
  fraction_cutoff = 5,
  source_cutoff = 3
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}

\item{wad_reference}{(\emph{tibble}) A tibble of WAD reference values}

\item{fraction_cutoff}{(\emph{numeric, default: 5}) The minimum number of fractions a feature must be found in to be considered}

\item{source_cutoff}{(\emph{numeric, default: 3}) The minimum number of sources a feature must be found in to be considered}
}
\description{
Make nested dataframe for gpd adjustment sampling
}
\keyword{internal}
