% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_validation.R
\name{isotope_validation}
\alias{isotope_validation}
\title{Check the validity of an isotope string}
\usage{
isotope_validation(
  isotope,
  isotope_list = c("12C", "13C", "14N", "15N", "16O", "18O")
)
}
\arguments{
\item{isotope}{A string or vector of isotope strings}

\item{isotope_list}{A master list of isotopes to check against}
}
\value{
A TRUE boolean if the isotope strings are valid, or a printed error

keywords internal
}
\description{
This is a helper function.
}
