% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgi_sample_df.R
\name{jgi_sample_df}
\alias{jgi_sample_df}
\title{Make a sample dataframe from JGI proposal file}
\usage{
jgi_sample_df(file_path, skip = 27)
}
\description{
Make a sample dataframe from JGI proposal file
}
