% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgi_source_df.R
\name{jgi_source_df}
\alias{jgi_source_df}
\title{Make a source dataframe from JGI proposal file}
\usage{
jgi_source_df(file_path, skip = 27)
}
\description{
Make a source dataframe from JGI proposal file
}
