% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_wrapper.R
\name{multi_qsip_wrapper}
\alias{multi_qsip_wrapper}
\title{Wrapper for running the qSIP EAF workflow}
\usage{
multi_qsip_wrapper(
  qsip_data_object,
  group = NULL,
  unlabeled_source_mat_ids,
  labeled_source_mat_ids,
  min_unlabeled_sources = 2,
  min_labeled_sources = 2,
  min_unlabeled_fractions = 2,
  min_labeled_fractions = 2,
  allow_failures = FALSE,
  resamples = 1000,
  seed = NULL
)
}
\description{
Wrapper for running the qSIP EAF workflow
}
