% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{multi_qsip_wrapper_launcher}
\alias{multi_qsip_wrapper_launcher}
\title{Internal function for launching multi_qsip_wrapper()}
\usage{
multi_qsip_wrapper_launcher(
  group,
  name,
  qsip_data_object,
  min_unlabeled_sources = 2,
  min_labeled_sources = 2,
  min_unlabeled_fractions = 2,
  min_labeled_fractions = 2,
  allow_failures = FALSE,
  resamples = 1000,
  seed = NULL
)
}
\arguments{
\item{group}{A group dataframe}

\item{name}{A group name}

\item{qsip_data_object}{A qsip_data object}

\item{min_unlabeled_sources}{The minimum number of sources for the unlabeled isotope}

\item{min_labeled_sources}{The minimum number of sources for the labeled isotope}

\item{min_unlabeled_fractions}{The minimum number of fractions for the unlabeled isotope}

\item{min_labeled_fractions}{The minimum number of fractions for the labeled isotope}

\item{allow_failures}{Whether to allow failures in the filtering step}

\item{resamples}{The number of resamples to run}

\item{seed}{The seed for the resampling}
}
\description{
Internal function for launching multi_qsip_wrapper()
}
\keyword{internal}
