% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_resamples.R
\name{n_resamples}
\alias{n_resamples}
\title{Number of resamples}
\usage{
n_resamples(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} object or list of objects}
}
\description{
This function returns the number of resamples that were performed on the data.
It will return \code{NULL} if the data has not been resampled.
}
