% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{plot_difference_to_mean}
\alias{plot_difference_to_mean}
\title{Plot difference between feature WAD and WAD reference}
\usage{
plot_difference_to_mean(
  qsip_data_object,
  fraction_cutoff = 5,
  source_cutoff = 3,
  quiet = F
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object}

\item{fraction_cutoff}{(\emph{numeric, default: 5}) The minimum number of fractions a feature must be found in to be considered}

\item{source_cutoff}{(\emph{numeric, default: 3}) The minimum number of sources a feature must be found in to be considered}

\item{quiet}{(\emph{logical, default: FALSE}) Whether to print a message about the number of features found}
}
\description{
Plot difference between feature WAD and WAD reference
}
