% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_curves.R
\name{plot_feature_curves}
\alias{plot_feature_curves}
\title{Plot qSIP feature data density curves}
\usage{
plot_feature_curves(qsip_data_object, feature_ids, source_mat_ids = NULL)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip data object with tube relative abundances}

\item{feature_ids}{(\emph{string}) Feature ids to be plotted on their own facet}

\item{source_mat_ids}{(\emph{string, defaults to all}) A list of source material ids}
}
\value{
A ggplot object
}
\description{
Plot qSIP feature data density curves
}
\seealso{
Other "Visualizations": 
\code{\link{plot_sample_curves}()}
}
\concept{"Visualizations"}
