% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_filter_efficiency}
\alias{plot_filter_efficiency}
\title{Efficiency scatter plot with labeled/unlabeled/retained categories}
\usage{
plot_filter_efficiency(
  qsip_data_object,
  ncol = NULL,
  nrow = NULL,
  use_counts = FALSE
)
}
\arguments{
\item{qsip_data_object}{A filtered qsip_data object (or list)}

\item{add_diagonal}{Add reference diagonal line}
}
\value{
A ggplot2 object
}
\description{
Efficiency scatter plot with labeled/unlabeled/retained categories
}
