% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_filter_means}
\alias{plot_filter_means}
\title{Plot features vs abundance with labeled/unlabeled/retained categories}
\usage{
plot_filter_means(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A filtered qsip_data object (or list)}
}
\value{
A ggplot2 object
}
\description{
Plot features vs abundance with labeled/unlabeled/retained categories
}
