% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_sample_object.R
\name{plot_sample_curves}
\alias{plot_sample_curves}
\title{Plot qSIP sample data density curves}
\usage{
plot_sample_curves(x, ...)
}
\arguments{
\item{x}{An object of \code{qsip_sample_object} class}
}
\description{
Plot qSIP sample data density curves
}
\keyword{sample_data}
