% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_curves.R
\name{plot_sample_curves}
\alias{plot_sample_curves}
\title{Plot qSIP sample data density curves}
\usage{
plot_sample_curves(sample_data, source_data = NULL, colors = NULL)
}
\arguments{
\item{sample_data}{(\emph{qsip_sample_data or qsip_data}) Sample data that holds density and abundance values}

\item{source_data}{(\emph{qsip_source_data}) Optional data required if \code{sample_data} is a \code{qsip_sample_data} object}
}
\value{
A ggplot object
}
\description{
Plot qSIP sample data density curves
}
\seealso{
Other "visualizations": 
\code{\link{plot_density_outliers}()},
\code{\link{plot_feature_curves}()},
\code{\link{plot_filter_gradient_position}()},
\code{\link{plot_source_wads}()}
}
\concept{"visualizations"}
