% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_curves.R
\name{plot_sample_curves}
\alias{plot_sample_curves}
\title{Plot qSIP sample data density curves}
\usage{
plot_sample_curves(
  qsip_data_object,
  title = NULL,
  facet_by = "source",
  show_wad = FALSE,
  colors = lifecycle::deprecated()
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) qSIP object}

\item{title}{(\emph{character}) An optional title for the plot}

\item{facet_by}{(\emph{character}) Facet the plots by "source" or by "isotope"}

\item{show_wad}{(\emph{logical}) A logical value}
}
\value{
A ggplot object
}
\description{
Plot qSIP sample data density curves
}
\seealso{
Other "visualizations": 
\code{\link{plot_density_outliers}()},
\code{\link{plot_feature_curves}()},
\code{\link{plot_filter_results}()},
\code{\link{plot_source_wads}()}
}
\concept{"visualizations"}
