% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_curves.R
\name{plot_sample_curves}
\alias{plot_sample_curves}
\title{Plot qSIP sample data density curves}
\usage{
plot_sample_curves(qsip_data_object, colors = NULL, title = NULL)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) qSIP object}

\item{colors}{(\emph{character}) A named vector of colors for each isotope (optional)}

\item{title}{(\emph{character}) An optional title for the plot}
}
\value{
A ggplot object
}
\description{
Plot qSIP sample data density curves
}
\seealso{
Other "visualizations": 
\code{\link{plot_density_outliers}()},
\code{\link{plot_feature_curves}()},
\code{\link{plot_filter_results}()},
\code{\link{plot_source_wads}()}
}
\concept{"visualizations"}
