% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_source_wads.R
\name{plot_source_wads}
\alias{plot_source_wads}
\title{Plot the source WADs by isotope}
\usage{
plot_source_wads(sample_data, source_data = NULL, group = NULL, colors = NULL)
}
\arguments{
\item{sample_data}{(\emph{qsip_data or qsip_sample_data}) A qsip object with sample data and optional source data}

\item{source_data}{(\emph{qsip_source_data, optional}) An optional qsip_source_data object. Required if \code{sample_data} does not contain the source_data}

\item{group}{(\emph{string}) An optional grouping parameter to facet the y or x,y axes}

\item{color}{(\emph{string}) An optional override to the default color palette}
}
\value{
A ggplot object
}
\description{
Plot the source WADs by isotope
}
\seealso{
Other "visualizations": 
\code{\link{plot_density_outliers}()},
\code{\link{plot_feature_curves}()},
\code{\link{plot_filter_gradient_position}()},
\code{\link{plot_sample_curves}()}
}
\concept{"visualizations"}
