% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_objects.R
\name{qsip_data}
\alias{qsip_data}
\title{qSIP master data class}
\usage{
qsip_data(source_data, sample_data, feature_data)
}
\value{
A validated \code{qsip_data} object
}
\description{
A class to hold and validate qSIP data.
}
\details{
This \code{qSIP2} object holds the source, sample and feature data, and does some
validation that the necessary source_mat_ids and sample_ids are correctly
shared.
}
\section{Slots}{

\describe{
\item{\code{source_data}}{(\emph{qsip_source_data}) A qSIP source data object}

\item{\code{sample_data}}{(\emph{qsip_sample_data}) A qSIP sample data object}

\item{\code{feature_data}}{(\emph{qsip_feature_data}) A qSIP feature data object}
}}

\seealso{
Other "qSIP Objects": 
\code{\link{qsip_feature_data}()},
\code{\link{qsip_sample_data}()},
\code{\link{qsip_source_data}()}
}
\concept{"qSIP Objects"}
