% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_feature_data.R
\name{qsip_feature_data}
\alias{qsip_feature_data}
\title{qSIP feature table class}
\usage{
qsip_feature_data(data, feature_id)
}
\value{
A validated object of the \code{qsip_feature_data} type
}
\description{
A class to hold and validate a feature abundance table
}
\section{Slots}{

\describe{
\item{\code{data}}{(\emph{dataframe}) ASV/OTU table or equivalent}

\item{\code{feature_id}}{(\emph{string}) Column name with unique taxa IDs}
}}

\seealso{
Other "qSIP Objects": 
\code{\link{qsip_data}()},
\code{\link{qsip_sample_data}()},
\code{\link{qsip_source_data}()}
}
\concept{"qSIP Objects"}
\keyword{object}
