% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_feature_data.R
\name{qsip_feature_data}
\alias{qsip_feature_data}
\title{qSIP feature table class}
\usage{
qsip_feature_data(data, feature_id)
}
\value{
A \code{qsip_feature_data} that holds a validated feature abundance table.
}
\description{
A class to hold and validate a feature abundance table
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe or tibble}

\item{\code{feature_id}}{Column name with unique taxa IDs}
}}

\seealso{
Other "qSIP classes": 
\code{\link{qsip_sample_data}()},
\code{\link{qsip_source_data}()}
}
\concept{"qSIP classes"}
\keyword{object}
