% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_sample_data.R
\name{qsip_sample_data}
\alias{qsip_sample_data}
\title{qSIP sample data class}
\usage{
qsip_sample_data(
  data,
  sample_id,
  source_mat_id,
  gradient_position,
  gradient_pos_density,
  gradient_pos_amt,
  gradient_pos_rel_amt
)
}
\description{
A class to hold and validate sample data.
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe or tibble}

\item{\code{sample_id}}{The unique ID for the sample/fraction}

\item{\code{source_mat_id}}{The unique ID for the biological subject or replicate}

\item{\code{gradient_position}}{Column name with the fraction position}

\item{\code{gradient_pos_density}}{Column name with the gradient density}

\item{\code{gradient_pos_amt}}{Column name with a total amount per fraction, either
qPCR copies or DNA}

\item{\code{gradient_pos_rel_amt}}{Column name with the relative fraction abundance
compared to the total}

\item{\code{fraction_volume}}{The volume loaded onto the column. Required if the \code{gradient_pos_amt} is reported as a concentration}
}}

\keyword{object}
