% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_sample_data.R
\name{qsip_sample_data}
\alias{qsip_sample_data}
\title{qSIP sample data class}
\usage{
qsip_sample_data(
  data,
  sample_id = "sample_id",
  source_mat_id = "source_mat_id",
  gradient_position = "gradient_position",
  gradient_pos_density = "gradient_pos_density",
  gradient_pos_amt = "gradient_pos_amt",
  gradient_pos_rel_amt = "gradient_pos_rel_amt"
)
}
\value{
A validated object of the \code{qsip_sample_data} type
}
\description{
A class to hold and validate sample data.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\emph{dataframe}) Metadata for samples/fractions}

\item{\code{sample_id}}{(\emph{string}) The unique sample ID}

\item{\code{source_mat_id}}{(\emph{string}) The unique ID for the biological subject or replicate}

\item{\code{gradient_position}}{(\emph{string}) Column name with the fraction position}

\item{\code{gradient_pos_density}}{(\emph{string}) Column name with the gradient density}

\item{\code{gradient_pos_amt}}{(\emph{string}) Column name with a total amount per fraction, either
qPCR copies or DNA}

\item{\code{gradient_pos_rel_amt}}{(\emph{string}) Column name with the relative fraction abundance
compared to the total}

\item{\code{fraction_volume}}{(\emph{string}) The volume loaded onto the column. Required if the \code{gradient_pos_amt} is reported as a concentration}
}}

\seealso{
Other "qSIP Objects": 
\code{\link{qsip_data}()},
\code{\link{qsip_feature_data}()},
\code{\link{qsip_source_data}()}
}
\concept{"qSIP Objects"}
