% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_sample_object.R
\name{qsip_sample_object}
\alias{qsip_sample_object}
\title{qSIP sample data class}
\usage{
qsip_sample_object(
  data = class_missing,
  isotope = class_missing,
  isotopolog_label = class_missing,
  gradient_position = class_missing,
  gradient_pos_density = class_missing,
  gradient_pos_amt = class_missing,
  gradient_pos_rel_amt = class_missing,
  source_mat_id = class_missing
)
}
\description{
A class to hold and validate sample data.
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe or tibble}

\item{\code{isotope}}{Column name with the MISIP isotope data}

\item{\code{isotopolog}}{Column name with the MISIP isotopolog data}
}}

