% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_source_data.R
\name{qsip_source_data}
\alias{qsip_source_data}
\title{qSIP source data class}
\usage{
qsip_source_data(data, isotope, isotopolog, source_mat_id)
}
\value{
A validated \code{qsip_source_data} object
}
\description{
A class to hold and validate qSIP source data.
}
\details{
This \code{qSIP2} object holds source data for calculating EAF values. If you want
to also calculate growth then you should use the \code{qsip_source_data_growth}
object.
}
\section{Slots}{

\describe{
\item{\code{data}}{(\emph{dataframe}) Source metadata}

\item{\code{isotope}}{(\emph{string}) Isotope name}

\item{\code{isotopolog}}{(\emph{string})  Isotopolog data}

\item{\code{source_mat_id}}{(\emph{string}) The unique ID for the biological subject or source}
}}

\seealso{
Other "qSIP classes": 
\code{\link{qsip_data}()},
\code{\link{qsip_feature_data}()},
\code{\link{qsip_sample_data}()}
}
\concept{"qSIP classes"}
