% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsip_source_data.R
\name{qsip_source_data}
\alias{qsip_source_data}
\title{qSIP source data class}
\usage{
qsip_source_data(data, isotope, isotopolog, source_mat_id)
}
\value{
A \code{qsip_source_data} object
}
\description{
A class to hold and validate qSIP source data.
}
\details{
This \code{qSIP2} object holds source data for calculating EAF values. If you want
to also calculate growth then you should use the \code{qsip_source_data_growth}
object.
}
\section{Slots}{

\describe{
\item{\code{data}}{Source metadata (dataframe)}

\item{\code{isotope}}{Isotope name (string)}

\item{\code{isotopolog}}{Isotopolog data (string)}

\item{\code{source_mat_id}}{The unique ID for the biological subject or source (string)}
}}

\keyword{object}
