% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_isotopolog_label.R
\name{remove_isotopolog_label}
\alias{remove_isotopolog_label}
\title{Remove isotopolog_label from data}
\usage{
remove_isotopolog_label(data)
}
\arguments{
\item{data}{A dataframe or tibble}
}
\description{
To convert from data with both isotope and isotopolog_label to one with isotope only
}
\examples{
remove_isotopolog_label(sample_data_MISIP)

}
\keyword{sample_data}
