% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISIP.R
\name{remove_isotopolog_label_check}
\alias{remove_isotopolog_label_check}
\title{Remove isotopolog label if needed}
\usage{
remove_isotopolog_label_check(df, isotope = "isotope")
}
\description{
Remove isotopolog label if needed
}
\concept{MISIP}
