% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_seed.R
\name{resample_seed}
\alias{resample_seed}
\title{Seed used in resampling}
\usage{
resample_seed(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} object that has been run through \code{run_resampling()}}
}
\description{
Returns the seed used in the resampling step, or \code{NA} if no specific seed was given.
}
