% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{resample_seed}
\alias{resample_seed}
\title{Seed used in resampling}
\usage{
resample_seed(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{A \code{qSIP_data} or list of \code{qSIP_data} objects.}
}
\description{
Returns the seed used in the resampling step, or \code{NULL} if no specific seed was given.
}
