% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EAF_calculations.R
\name{run_EAF_calculations}
\alias{run_EAF_calculations}
\title{Calculate EAF values}
\usage{
run_EAF_calculations(qsip_data_object)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) A qsip_data_object with resample information}
}
\value{
Returns an updated \code{qsip_data_object} with final EAF and other values
in the \verb{@EAF} slot.
}
\description{
This is the last main workhorse function in qSIP analysis and generates the important
values (Z, G, M, M_labeledmax, M_labeled and EAF) for the observed data as well
as for all of the resampled data.
}
