% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_growth_calculations}
\alias{run_growth_calculations}
\title{Run growth calculations}
\usage{
run_growth_calculations(
  qsip_data_object,
  N_total_it,
  growth_model = "exponential",
  timepoint = "timepoint",
  correct_copy_numbers = "filter",
  correct_EAF = "filter",
  propO = 1
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{N_total_it}{(\emph{data.frame}) A data frame of time zero totals from \code{get_N_total_it()}}

\item{growth_model}{(\emph{character, default: exponential}) The growth model to use. Must be either "exponential" or "linear"}

\item{timepoint}{(\emph{character}) The name of the timepoint column in the source data}

\item{correct_copy_numbers}{(\emph{character, default: filter}) If copy numbers are not logical (e.g. < 0), should they be filtered out or adjusted to 0?}

\item{correct_EAF}{(\emph{character, default: filter}) If EAF values are not logical (e.g. <0 or >1), should they be filtered out or adjusted to 0 or 1?}

\item{propO}{(\emph{numeric}) The proportion of heavy isotope in the labeled DNA. Only used for 18O.}
}
\description{
Run growth calculations
}
