% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_growth_calculations.R
\name{run_growth_calculations}
\alias{run_growth_calculations}
\title{Run growth calculations}
\usage{
run_growth_calculations(
  qsip_data_object,
  time_zero_totals,
  timepoint = "timepoint"
)
}
\arguments{
\item{qsip_data_object}{(\emph{qsip_data}) An object of \code{qsip_data} class}

\item{time_zero_totals}{(\emph{data.frame}) A data frame of time zero totals from \code{calculate_time_zero_totals()}}
}
\description{
Run growth calculations
}
