% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_comparison_groups.R
\name{show_comparison_groups}
\alias{show_comparison_groups}
\title{Show comparison groups}
\usage{
show_comparison_groups(
  data,
  tmt,
  isotope = "isotope",
  source_mat_id = "source_mat_id"
)
}
\arguments{
\item{data}{Sample metadata (dataframe)}

\item{tmt}{Treatment value or values (string(s))}

\item{isotope}{Column name with isotope data (string, default: "isotope")}

\item{source_mat_id}{Column name with source_mat_id (string, default: "source_mat_id")}
}
\value{
A dataframe with id grouped by different \code{tmt} treatments and isotopes
}
\description{
Generates a table of ids grouped in columns by isotope, and in rows by the
given treatments.
}
\keyword{sample_data}
