% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{show_comparison_groups}
\alias{show_comparison_groups}
\title{Show comparison groups (deprecated)}
\usage{
show_comparison_groups(
  source_data = NULL,
  group = NULL,
  isotope = "isotope",
  source_mat_id = "source_mat_id"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Generates a table of ids grouped in columns by isotope, and in rows by the
given treatments.
}
