% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_comparison_groups.R
\name{show_comparison_groups}
\alias{show_comparison_groups}
\title{Show comparison groups}
\usage{
show_comparison_groups(
  source_data = NULL,
  group = NULL,
  isotope = "isotope",
  source_mat_id = "source_mat_id"
)
}
\arguments{
\item{source_data}{(\emph{dataframe, qsip_source_data or qsip_data}) Sample metadata}

\item{group}{(\emph{string}) Treatment value or values}

\item{isotope}{(\emph{string, default: isotope}) Column name with isotope data}

\item{source_mat_id}{(\emph{string, default: source_mat_id}) Column name with source_mat_id}
}
\value{
A dataframe with id grouped by different \code{group} treatments and isotopes
}
\description{
Generates a table of ids grouped in columns by isotope, and in rows by the
given treatments.
}
