% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abundances.R
\name{validate_abundances}
\alias{validate_abundances}
\title{Check the validity of a feature abundance table (internal)}
\usage{
validate_abundances(data, feature_id, type)
}
\arguments{
\item{data}{(\emph{dataframe}) ASV/OTU table}

\item{feature_id}{(\emph{string}) Column name with feature IDs}

\item{type}{(\emph{string}) \emph{counts} requires integers, \emph{coverage} and \emph{relative} can take any positive numeric}
}
\value{
Returns \code{NULL} if the values are valid, or a printed error
}
\description{
This validation function is an internal function will check the validity of a
feature abundance table.
}
\details{
Rows should contain the unique taxa ids with a column designated with the id
argument. Each other column name should be a unique sample name.

Validity checking includes making sure all data is numeric (except for the
feature IDs), all numbers are integers (if \code{type = "counts"}), and no numbers are
negative.
}
