% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_isotopes.R
\name{validate_isotopes}
\alias{validate_isotopes}
\title{Check the validity of an isotope string}
\usage{
validate_isotopes(
  isotope,
  isotope_list = c("12C", "13C", "14N", "15N", "16O", "18O")
)
}
\arguments{
\item{isotope}{(\emph{string(s)}) Isotope value or values}

\item{isotope_list}{(\emph{strings, default: c("12C", "13C", "14N", "15N", "16O", "18O")}) Isotopes to check against}
}
\value{
Returns \code{NULL} if the isotope strings are valid, or a printed error
}
\description{
Check the validity of an isotope string
}
\note{
The isotope_list may change if isotopolog_label stays a thing. Only the "labeled" isotopes will be allowed.
}
