% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_isotopes}
\alias{validate_isotopes}
\title{Check the validity of an isotope string (internal)}
\usage{
validate_isotopes(
  isotope,
  isotope_list = c(valid_isotope_names$unlabeled, valid_isotope_names$labeled),
  unfractionated_terms = valid_isotope_names$unfractionated
)
}
\arguments{
\item{isotope}{(\emph{string(s)}) Isotope value or values}

\item{isotope_list}{(\emph{strings, default: c("12C", "13C", "14N", "15N", "16O", "18O")}) Isotopes to check against}

\item{unfractionated_terms}{(\emph{strings}) Terms to ignore when checking isotope values}
}
\value{
Returns \code{NULL} if the isotope strings are valid, or a printed error
}
\description{
Often the "bulk" designation is found in this column, so those can optionally
be removed from validation checks
}
\note{
The isotope_list may change if isotopolog_label stays a thing. Only the "labeled" isotopes will be allowed.
}
\keyword{internal}
