% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_source_isotope}
\alias{validate_source_isotope}
\title{Validate the given source mat ids have the expected labeled/unlabeled designation (internal)}
\usage{
validate_source_isotope(qsip_data_object, source_mat_ids, isotope_list)
}
\arguments{
\item{qsip_data_object}{A qsip_data object}

\item{source_mat_ids}{A character vector of source_mat_ids}

\item{isotope_list}{A character vector of isotopes to check against}
}
\value{
TRUE (all match) or FALSE (some don't match)
}
\description{
Currently used by \code{run_feature_filter()} to make sure user given labeled or unlabeled
source_mat_ids are not incorrect with respective to their source_data.
}
\keyword{internal}
